+++ src/migrate.ts
@@ -54,7 +54,7 @@ export function migrateZodV3ToV4(
 
     const parentType = node.getFirstAncestorByKind(SyntaxKind.QualifiedName);
     if (parentType?.getText().endsWith("ZodSchema")) {
-      parentType?.getRight().replaceWithText("ZodJSONSchema");
+      parentType?.getRight().replaceWithText("ZodType");
     }
 
     const parentStatement = node.getParentWhile(isZodNode) || node;