+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- Converts `ZodSchema` into `ZodType` instead of `ZodJSONSchema`, which was a mistake ([#28](https://github.com/nicoespeon/zod-v3-to-v4/pull/28))
+
 ## [1.7.0] - 2025-11-05
 
 ### Added